Recursive Call should always succeed
We call Wait from the outer goroutines, and expect that all submitted
items will be processed.  However, some of those items may have enqueued
other items to be processed, and it would be incorrect to not include
those in the result.  This commit fixes that by ensuring that Call always
succeeds when being called from inside Integrate.