Perform some refactoring
This commit contains a fair amount of refactoring, including changing the
name of the ErrWorkerClosed error to ErrClosed; renaming the worker
states; exposing the runResult struct as Result; and refactoring panicer
to call a function and return a Result.  The interface to the
Runner.Integrate method is also updated to take a pointer to a Result and
to add some probably necessary locks within parallelWorker.getResult.