Add a serializer
This commit creates a serializer interface and implementation.  A
serializer is the opposite of a parallelizer, in that the calls are made
synchronously in a single goroutine, instead of in parallel by a farm of
worker goroutines.  This may be used to allow multiple goroutines to
invoke methods on an object which is not designed with thread safety in
mind; in particular, the interface is perfect for using with a builder
pattern.