Fix parallel manager
The parallel manager was exiting when all the work had been sent to the
workers, but before it had received results from the workers; and some of
those workers were sending back results that added new work that now there
weren't any workers to work.  This commit adds a "waiting" field to the
parallelManager struct that counts the number of results we're waiting on,
and the "shut down workers" logic now also demands that that value be less
than or equal to 0.  This solves the deadlock problem.