Add tool installation
This commit adds code to the Makefile to on-demand install the tools that
are required for the make operation.  The tools are installed into a new
tools directory with its own go.mod to avoid dirtying up the package's
go.mod.  A new (explicitly undocumented) goveralls target is also added,
to leverage this tooling for goveralls in the Travis run.  The .travis.yml
is also updated to leverage the new tooling; no install section is needed
anymore.